fix: correct user_id data type in contacts table migration
- Changed user_id from BIGINT UNSIGNED to INT UNSIGNED to match users.id
- This fixes MySQL foreign key constraint incompatibility error

+++ internal/app/migrations/migrations.go
@@ -48,7 +48,7 @@ func GetMigrations() []Migration {
 				_, err := tx.Exec(`
 					CREATE TABLE IF NOT EXISTS contacts (
 						id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-						user_id BIGINT UNSIGNED NOT NULL,
+						user_id INT UNSIGNED NOT NULL,
 						full_name VARCHAR(255) NOT NULL,
 						phone VARCHAR(20) NOT NULL,
 						email VARCHAR(255) NULL,
